requirements.txt
streamlit
pandas
langchain
openpyxl
python-dotenv